Added a dialogue box before deleting a user

+++ frontend/src/screens/UserListScreen.js
@@ -32,7 +32,9 @@ const UserListScreen = ({ history }) => {
 	);
 
 	const deleteHandler = (id) => {
-		dispatch(deleteUser(id));
+		if (window.confirm('Are you sure?')) {
+			dispatch(deleteUser(id));
+		}
 	};
 
 	return (
